refactor(usage): split usage-cost helpers and pricing provenance tests under max-lines
